+++ HISTORY.md
@@ -2,7 +2,7 @@
 https://github.com/enmasseio/babble
 
 
-## not yet released, version 0.8.1
+## 2014-08-11, version 0.9.0
 
 - Implemented `'default'` choice in `Decision`.
 - Some code simplifications.